Change type from int to double

+++ src/com/example/Math.java
@@ -6,7 +6,7 @@ public class Math {
         return a + b;
     }
 
-    public int multiply(int a, int b) {
+    public double multiply(double a, double b) {
         return a * b;
     }
 }
